pyproject.toml
[project]
name = "llm-query-transport"
version = "0.1.0"
description = ""
authors = [{ name = "Juan Pineda", email = "juandpineda@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas>=2.0.3",
    "numpy>=1.25.1",
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "psycopg2-binary>=2.9.7",
    "scikit-learn>=1.3.0",
    "keplergl>=0.3.2",
    "pydeck>=0.8.0",
    "sqlalchemy>=2.0.25",
    "openai>=1.12.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.109.2",
    "pyprojroot>=0.3.0",
    "streamlit>=1.31.1",
    "streamlit-aggrid>=1.0.5",
    "streamlit-tags>=1.2.8",
    "streamlit-autorefresh>=1.0.1",
    "faiss-cpu>=1.7.2",
    "requests-oauthlib>=1.3.0",
    "langchain>=0.2.13,<1.0",
    "langchain-openai>=0.1.21,<1.0",
    "langchain-core>=0.2.30,<1.0",
    "langchain-community>=0.2.12,<1.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.12.0",
    "scipy~=1.11.0",
    "watchdog>=3.0.0,<6.0.0",
    "structlog>=25.4.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.24.0",
]

[tool.mypy]
ignore_missing_imports = true
exclude = [
    "notebooks/.*",
    ".*\\.ipynb$",
    "experiments/.*",
    "scripts/.*",
    "__pycache__/.*"
]

[tool.pyright]
reportMissingImports = false
reportMissingTypeStubs = false
reportUnboundVariable = false
reportMissingModuleSource = false
reportUnusedFunction = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

# Default ruff configuration
[tool.ruff]
line-length = 140
extend-exclude = [".venv", "libs", "out", ".git", ".mypy_cache", ".tox", "venv", "_build", "buck-out", "build", "dist"]

[tool.ruff.lint]
select = ["ALL"]
extend-ignore = [
    "D211", "D212", "D213", "D100", "TD002", "TD003", "D401", "FA100", "FA102",
    "D104", "D101", "D102", "D103", "D107", "D205", "D200", "D202",
    "ANN001", "ANN201", "ANN202", "ANN204", "ANN401",
    "FBT001", "FBT002", "PLR0913", "B008", "ARG001", "ARG002",
    "BLE001", "PLR2004", "TRY003", "EM101", "EM102", "G004",
    "T201", "ERA001", "FIX002", "PTH", "TRY002", "TRY400", "TRY401",
    "S113", "PLR0911", "PLR0912", "C901", "INP001", "PLC0415",
    "RET501", "RET504", "RET505", "PT006", "PT011", "PT012",
    "B017", "FAST002", "D415", "PLR0402", "E501", "RUF002",
    "TRY300", "N806", "PLR0915",
    "D203", "COM812",
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN201"]
"experiments/*" = ["T201"]
"src/config/settings.py" = ["E402"]
